requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
pydantic==2.12.5
python-dotenv==1.2.2
langchain-openai==1.1.10
langchain-core==1.2.17
langgraph==1.0.10
pandas==2.3.3
typing-extensions==4.15.0
psycopg2-binary==2.9.11
pymysql==1.1.1
